app/requirements.txt
Django
django-cors-headers
djangorestframework
psycopg2-binary
python-dotenv
urllib3
